Add LineStatus model & network service implementation
LineStatus class represents the data described in XML from
http://web.mta.info/status/serviceStatus.txt for the status
of a single subway line. Other modes of transportation are
included too (buses, Metro North, LIRR) but are ignored for now.

The network service retrieves this XML data and parses it into
LineStatus objects. It uses MtaStatusXmlParser (which uses XmlPullParser)
to accomplish the parsing.