Add notification feature to background job
The MtaAlertJobService launches a notification showing the data it
just retrieved. Also, the JobScheduler was updated to use a combo of
setMinimumLatency() and setOverrideDeadline() rather than setPeriodic()
because the latter does not allow any delay before the initial execution
of a periodic job, and we don't want a notification to launch right away
after the app runs.